Adding ci server
