+++ src/main/response.json
@@ -1,0 +1,1 @@
+[{'id': 814570189, 'iid': 1512608, 'project_id': 278964, 'sha': '547c8358920ca17162bfd923d7c9de9d6c4dea9b', 'ref': 'refs/merge-requests/115528/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:33:47.556Z', 'updated_at': '2023-03-22T14:33:54.303Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814570189'}, {'id': 814569400, 'iid': 1512607, 'project_id': 278964, 'sha': 'cc09004666c6ce34a4a71eed283577561b316c08', 'ref': 'refs/merge-requests/115537/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:33:20.943Z', 'updated_at': '2023-03-22T14:33:26.977Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814569400'}, {'id': 814567808, 'iid': 1512605, 'project_id': 278964, 'sha': 'e7576c51416d96ccded42e78e72e06701b572094', 'ref': 'refs/merge-requests/115059/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:32:27.171Z', 'updated_at': '2023-03-22T14:32:32.928Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814567808'}, {'id': 814566829, 'iid': 1512603, 'project_id': 278964, 'sha': 'ae6400b6dd2c24d449a425d7d5dddf65a56e7c1c', 'ref': 'refs/merge-requests/115536/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:31:52.653Z', 'updated_at': '2023-03-22T14:32:00.096Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814566829'}, {'id': 814566685, 'iid': 1512602, 'project_id': 278964, 'sha': '3bcd04b0fc9dd01ee58bcc4413b34298070b3d9b', 'ref': 'refs/merge-requests/115535/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:31:47.733Z', 'updated_at': '2023-03-22T14:31:48.868Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814566685'}, {'id': 814564585, 'iid': 1512601, 'project_id': 278964, 'sha': 'c5c894103d818565c15fb00371560d4ed2c83776', 'ref': 'refs/merge-requests/113937/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:30:42.294Z', 'updated_at': '2023-03-22T14:30:48.413Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814564585'}, {'id': 814561148, 'iid': 1512596, 'project_id': 278964, 'sha': '0f9e821fe677b0e4bafebbf8f3d66d9cb0dd2d4f', 'ref': 'refs/merge-requests/113040/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:28:59.991Z', 'updated_at': '2023-03-22T14:29:11.828Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814561148'}, {'id': 814560293, 'iid': 1512594, 'project_id': 278964, 'sha': '470b4066fc3e58004a7806cbf9c2751ae6f0e513', 'ref': 'refs/merge-requests/115445/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:28:24.356Z', 'updated_at': '2023-03-22T14:28:29.375Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814560293'}, {'id': 814557894, 'iid': 1512592, 'project_id': 278964, 'sha': 'cf7954b449213f69e5d26e5cd47ee1434c67a3bc', 'ref': 'refs/merge-requests/115534/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:26:53.643Z', 'updated_at': '2023-03-22T14:26:59.268Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814557894'}, {'id': 814556727, 'iid': 1512591, 'project_id': 278964, 'sha': '170fb5bb014b7685523a855add3c8d981a86a7f0', 'ref': 'refs/merge-requests/115075/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:25:52.910Z', 'updated_at': '2023-03-22T14:25:59.251Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814556727'}, {'id': 814555477, 'iid': 1512590, 'project_id': 278964, 'sha': '26562017b5a0f2f94930940b6419d596c4a42ba9', 'ref': 'refs/merge-requests/115523/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:24:57.718Z', 'updated_at': '2023-03-22T14:25:02.823Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814555477'}, {'id': 814554487, 'iid': 1512589, 'project_id': 278964, 'sha': 'ae1fb60125ec600781f87bc66c21d586b0778ba8', 'ref': 'refs/merge-requests/115402/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:24:09.832Z', 'updated_at': '2023-03-22T14:24:15.106Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814554487'}, {'id': 814553837, 'iid': 1512588, 'project_id': 278964, 'sha': '8b5fc0b7454ca8e579b953a56f29ed3dde821b4f', 'ref': 'philipcunningham-enhance-dependency-equality-390027', 'status': 'skipped', 'source': 'push', 'created_at': '2023-03-22T14:23:36.849Z', 'updated_at': '2023-03-22T14:23:36.849Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814553837'}, {'id': 814552605, 'iid': 1512587, 'project_id': 278964, 'sha': '6f7558f1418f824d445ee04844d86478c0c796fc', 'ref': 'master', 'status': 'running', 'source': 'push', 'created_at': '2023-03-22T14:22:36.229Z', 'updated_at': '2023-03-22T14:22:47.484Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814552605'}, {'id': 814549688, 'iid': 1512584, 'project_id': 278964, 'sha': 'e64d343e44481844c3aea36f0761df011548ada4', 'ref': 'refs/merge-requests/115430/merge', 'status': 'success', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:20:10.274Z', 'updated_at': '2023-03-22T14:29:14.384Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814549688'}, {'id': 814547119, 'iid': 1512580, 'project_id': 278964, 'sha': 'afac887a08ca0a803fc1b1e813487a237c62b550', 'ref': 'refs/merge-requests/115250/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:18:20.730Z', 'updated_at': '2023-03-22T14:18:26.216Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814547119'}, {'id': 814546753, 'iid': 1512578, 'project_id': 278964, 'sha': 'a0ec39682c17455c3e3431e9aea265535fcc595e', 'ref': 'master', 'status': 'running', 'source': 'push', 'created_at': '2023-03-22T14:18:02.967Z', 'updated_at': '2023-03-22T14:18:10.039Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814546753'}, {'id': 814546641, 'iid': 1512579, 'project_id': 278964, 'sha': '6abc6578f70d33c668051c45b7a9b90764b764b6', 'ref': 'refs/merge-requests/115532/merge', 'status': 'success', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:17:57.609Z', 'updated_at': '2023-03-22T14:27:04.946Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814546641'}, {'id': 814545783, 'iid': 1512575, 'project_id': 278964, 'sha': '16e57cdd16726b345a4e666c56813914dd44c67c', 'ref': 'refs/merge-requests/115533/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:17:19.639Z', 'updated_at': '2023-03-22T14:17:24.591Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814545783'}, {'id': 814542355, 'iid': 1512572, 'project_id': 278964, 'sha': '326d0f4eb01fb200e8ee4b65c23f1ddff6c447fe', 'ref': 'refs/merge-requests/115484/merge', 'status': 'running', 'source': 'merge_request_event', 'created_at': '2023-03-22T14:14:54.996Z', 'updated_at': '2023-03-22T14:15:04.146Z', 'web_url': 'https://gitlab.com/gitlab-org/gitlab/-/pipelines/814542355'}]
